Update app.py
add the output to a table with results from all experiments. First, we need to add two stateful variables as keys of st.session_state. The session state is preserved over new runs of the Streamlit application. Then, we add collecting results of experiments in the dataframe kept as st.session_state['df_experiment_results']